changed amount data type from double to BigDecimal

+++ gateway-api/src/main/java/com/gateway/service/PaymentLoggingService.java
@@ -1,5 +1,6 @@
 package com.gateway.service;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import com.gateway.utils.Constants;
@@ -8,7 +9,7 @@ public interface PaymentLoggingService {
 
 	public void paymentLog(Date requestDatetime, Date responsetDatetime, String rrn, String stan, String responseCode,
 			String responseDescription, String cnic, String mobile, String name, String consumerNumber, String billerId,
-			double amount, double charges, String activity, String paymentRefNo, String billerNumber,
+			BigDecimal amount, double charges, String activity, String paymentRefNo, String billerNumber,
 			String transactionStatus, String address, double transactionFees, double taxAmount, double total, String channel,
 			String billStatus, String tranDate, String tranTime, String province,String tranAuthId) throws Exception;
